recommendation-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
kafka-python==2.0.2
psycopg2-binary==2.9.9
redis==5.0.1
numpy==1.24.3
scikit-learn==1.3.2
requests==2.31.0